+++ src/pages/chaincodevulnerability-detection/chaincodevulnerability-detection.tsx
@@ -3,13 +3,14 @@
  * @Author: didadida262
  * @Date: 2024-08-29 10:18:39
  * @LastEditors: didadida262
- * @LastEditTime: 2024-08-30 15:06:53
+ * @LastEditTime: 2024-09-02 15:21:33
  */
 /* eslint-disable @typescript-eslint/restrict-template-expressions */
 /* eslint-disable prettier/prettier */
 
 import cn from "classnames";
 import { useEffect, useState } from "react";
+import { useNavigate } from "react-router";
 
 import IdentityInferenceDialogTitle from "@/assets/IdentityInferenceDialogTitle.png";
 import { ButtonCommonV2, EButtonType } from "@/components/ButtonCommonV2";
@@ -27,88 +28,50 @@ import { columnsChainCode } from "@/services/columns";
 import pattern from "@/styles/pattern";
 
 export function ChaincodevulnerabilityDetection() {
-  const [pageState, setPageState] = useState("search");
   const [value, setValue] = useState<any>(null);
   const [dataList, setDateList] = useState<any>([]);
+  const navigate = useNavigate();
 
   const startSearch = () => {
     // 开始查询
-    setPageState("res");
+    navigate("/threat-evidence/chaincodevulnerability-detection/result");
   };
-  useEffect(() => {
-    // 请求
-    const res1 = [
-      {
-        file: 1,
-        time: "2024-8-20 16:00",
-        num: 100,
-        type: "a",
-        resaon: "在Golang中，开发者可以通过将返回值赋给“_”变量来跳过接收返回值，如果返回值是Error类型，则可能会忽略错误。"
-      }
-    ];
-
-    setDateList([...res1, ...res1, ...res1, ...res1]);
-  }, []);
 
   return (
     <div className={cn(" w-full h-full pt-[0px] fadeIn", `${pattern.flexbet}`)}>
-      {pageState === "search"
-        ? <div className={cn(`w-full h-full ${pattern.flexCenter} `)}>
-            <div className={cn(` w-[662px] h-[523px] relative`)}>
-              <div
-                className={cn(
-                  `absolute top-0 left-0 w-full h-[54px] bg-[url('./assets/ChaincodevulnerabilityDetectionDialogTitle.png')] bg-cover bg-center`
-                )}
-              />
-              <div
-                className={cn(
-                  `pt-[66px] px-[106px] pb-[40px] absolute top-[54px] left-0 w-full h-[calc(100%_-_54px)] `,
-                  "bg-[#003F7A4D]",
-                  "border-solid border-[1px] border-[#D3EAFF] border-t-0 border-r-0 border-l-0",
-                  `flex flex-col gap-y-6 items-end `
-                )}
-              >
-                <InputCommonV4
-                  placeholder="输入链码信息"
-                  onInput={(val: any) => {
-                    setValue(val);
-                  }}
-                  className="w-full h-[301px]"
-                />
-
-                <ButtonCommonV2
-                  onClick={() => {
-                    startSearch();
-                  }}
-                >
-                  <span className="text-[#FFFFFF] text-[16px]">检测</span>
-                </ButtonCommonV2>
-              </div>
-            </div>
-          </div>
-        : <div className={cn(`w-full h-full flex flex-col `)}>
-            <div
-              className={cn(
-                "w-full h-[50px] flex items-center justify-start mb-1"
-              )}
-            >
-              <ResultComponentV2 title="检测报告" className="w-full h-full" />
-            </div>
-            <div
-              className={cn(
-                `w-full h-[348px] mb-5`,
-                `border-[2px] border-solid border-[#0D53B7]`,
-                `bg-[#0095FF66]`
-              )}
+      <div className={cn(`w-full h-full ${pattern.flexCenter} `)}>
+        <div className={cn(` w-[662px] h-[523px] relative`)}>
+          <div
+            className={cn(
+              `absolute top-0 left-0 w-full h-[54px] bg-[url('./assets/ChaincodevulnerabilityDetectionDialogTitle.png')] bg-cover bg-center`
+            )}
+          />
+          <div
+            className={cn(
+              `pt-[66px] px-[106px] pb-[40px] absolute top-[54px] left-0 w-full h-[calc(100%_-_54px)] `,
+              "bg-[#003F7A4D]",
+              "border-solid border-[1px] border-[#D3EAFF] border-t-0 border-r-0 border-l-0",
+              `flex flex-col gap-y-6 items-end `
+            )}
+          >
+            <InputCommonV4
+              placeholder="输入链码信息"
+              onInput={(val: any) => {
+                setValue(val);
+              }}
+              className="w-full h-[301px]"
             />
-            <div className={cn(` w-full h-[272px] mt-4`)}>
-              <TableCommonV2
-                className=""
-                data={dataList}
-                columns={columnsChainCode}
-              />
-            </div>
-          </div>}
+
+            <ButtonCommonV2
+              onClick={() => {
+                startSearch();
+              }}
+            >
+              <span className="text-[#FFFFFF] text-[16px]">检测</span>
+            </ButtonCommonV2>
+          </div>
+        </div>
+      </div>
     </div>
   );
 }
