+++ src/pages/chaincodevulnerability-detection/chaincodevulnerability-detectionResult.tsx
@@ -1,0 +1,73 @@
+/*
+ * @Description:
+ * @Author: didadida262
+ * @Date: 2024-08-29 10:18:39
+ * @LastEditors: didadida262
+ * @LastEditTime: 2024-09-02 15:20:32
+ */
+/* eslint-disable @typescript-eslint/restrict-template-expressions */
+/* eslint-disable prettier/prettier */
+
+import cn from "classnames";
+import { useEffect, useState } from "react";
+
+import IdentityInferenceDialogTitle from "@/assets/IdentityInferenceDialogTitle.png";
+import { ButtonCommonV2, EButtonType } from "@/components/ButtonCommonV2";
+import { InputCommonV4 } from "@/components/InputCommonV4";
+import { ResultComponent } from "@/components/ResultComponent";
+import { ResultComponentV2 } from "@/components/ResultComponentV2";
+import {
+  ISelectorItemProps,
+  SelectorCommonV2
+} from "@/components/SelectorCommonV2";
+import { TableCommonV2 } from "@/components/TableCommonV2";
+import { TableCommonV3 } from "@/components/TableCommonV3";
+import { TagComponent } from "@/components/TagComponent";
+import { columnsChainCode } from "@/services/columns";
+import pattern from "@/styles/pattern";
+
+export function ChaincodevulnerabilityDetectionResult() {
+  const [value, setValue] = useState<any>(null);
+  const [dataList, setDateList] = useState<any>([]);
+
+  useEffect(() => {
+    // 请求
+    const res1 = [
+      {
+        file: 1,
+        time: "2024-8-20 16:00",
+        num: 100,
+        type: "a",
+        resaon: "在Golang中，开发者可以通过将返回值赋给“_”变量来跳过接收返回值，如果返回值是Error类型，则可能会忽略错误。"
+      }
+    ];
+
+    setDateList([...res1, ...res1, ...res1, ...res1]);
+  }, []);
+
+  return (
+    <div className={cn(" w-full h-full pt-[0px] fadeIn", `${pattern.flexbet}`)}>
+      <div className={cn(`w-full h-full flex flex-col `)}>
+        <div
+          className={cn("w-full h-[50px] flex items-center justify-start mb-1")}
+        >
+          <ResultComponentV2 title="检测报告" className="w-full h-full" />
+        </div>
+        <div
+          className={cn(
+            `w-full h-[348px] mb-5`,
+            `border-[2px] border-solid border-[#0D53B7]`,
+            `bg-[#0095FF66]`
+          )}
+        />
+        <div className={cn(` w-full h-[272px] mt-4`)}>
+          <TableCommonV2
+            className=""
+            data={dataList}
+            columns={columnsChainCode}
+          />
+        </div>
+      </div>
+    </div>
+  );
+}
